import math p,r,t=map(int,input().split()) ci=p*(pow((1+r/100),t)) print('%0.2f'%ci)

+++ Compound_interest.py
@@ -1,3 +1,4 @@
-p,r,t=map(float,input().split())
-ci=p*(pow(1+(r/100),t))
-print("%.2f"%ci)
+import math
+p,r,t=map(int,input().split())
+ci=p*(pow((1+r/100),t))
+print('%0.2f'%ci)
